Refactor Player class
Player class was not actually modeling a Player but its boards' values. So it has been decided to refactor it naming it `PlayerBoard` so it represent the part of a Game Board that belongs to a `Player`.

A new type has been created to represent a `Player` and since a Game has only two Player and a Player (at least by now) holds no state at all it has been decided to model it as an `enum` with two value: `PLAYER_1` and `PLAYER_2`.